requirements.txt
DearPyGui-DragAndDrop
dearpygui
annotated-types
numpy
pillow
